Corrected key reuse (and saved quite a bit of code)
The key that was used to authenticate the response was then used to
encrypt the sender's public key.  We've seen that Poly1305 renders its
authentication key useless for any purpose, so we can't have that.

The mechanics of key derivations have been streamlined.  Each time we
perform a key exchange, we derive 3 keys:

- A chaining key, derived from the previous chaining key and the current
  X22519 exchange.
- An authentication key, derived from the new chaining key.
- An encryption key, derived from the new chaining key.

When authenticating, we use the current authentication key.  This allows
one authentication per chaining key.

When encrypting 32 bytes, we XOR the current encryption key with the
input.  This allows up to one encryption per chaining key. (The last
encryption key is the session key, so we can't use it at all.)