Typo blunder in the specs
As it was, the spec was a bit pointless. We don't want to authenticate
the sender's key twice, that would be stupid. I meant to start the
transcript with the *recipient's* key (LR), to prevent an attack where
different binary representations of the same public key could cause some
confusion. (It's also good practice to authenticate everything just to
be sure we don't screw up.)