ER should not be encrypted after all
Reverts #2706605

Encrypting ER was a mistake. While it does make the protocol itself less
traceable, it also compromises the forward secrecy of the server's
identity for passive attackers. It goes like this:

1. Alice sends Es to Bob (and Eve)
2. Bob sends Er to Alice (and Eve), Encrypted with <el>
3. Some time later, Eve steals Bob's long term key lr.
4. Eve Tries to decrypt Er with every compromised key she has.
5. When trying Bob's lr, the result is a point on the curve.  Bingo.

Hiding the traffic is nice, but it costs us anonymity, *and* simplicity
*and* speed.  Better stick with the XK1 structure from Noise, whose
properties are well reviewed.

Users who want perfect untraceability will have to hide ES and ER with
Elligator. It's still possible to do on top of the current API, though
more expensive than it would have been if we encrypted ER.