Hash shared secret with an XChacha20 scheme
The previous entropy extractor, while cheaper, was easier to
mis-implement (leading to XOR catastrophes), and simply did not look as
safe.

This one seems to be the safest.  This is the same as XChacha20, except
without the final addition (which is not needed because we don't reveal
the whole Chacha20 block). The XChacha20 structure looks very safe
because XChacha20 proper is proven safe if Chacha20 is safe.

This is not XChacha20 proper, but we can conjecture that it doesn't
matter where the random material is (key, nonce, or counter), it will
get mixed just as well.  I feel pretty safe in assuming that a
quasi-random key can indeed be absorbed in two stages, 16 bits a time.