Gave up and used Blake2b
While Chacha20 would have been an elegant solution, I couldn't ensure
its safety.  The only way to have an easy to use, "no questions asked"
handshake is to use a primitive that was designed for this kind of
thing.

The effect on the runtime is barely measurable.

I've studied HKDF, but it causes a 2% speed hit, and requires more code.
Nope.  Naive keyed Blake2b is just as secure, there's no need for the
nuclear option.  (Besides, I've kept the "extract/expand" paradigm.)

We could have kept Chacha20 for the expand phase of the key derivation,
but Blake2b turned out to be just as fast, and ditching a primitive
makes the protocol a bit simpler.  (Encryption is done with XOR, which
is possible because the encrypted messages are short).

Authentication is still done with Poly1305.  We could use Blake2b, but
it's a tiny bit slower, and Poly1305 has a nice security reduction.
Besides, AEAD in Monocypher uses Poly1305, we might as well be
consistent with that choice.