Made initialisation explicit (breaks the API)
The previous API had too many arguments per function. By separating the
initialisation from the actual message processing, we reduce the number
of arguments, and make each function a bit easier to handle (the user is
less likely to swap arguments by accident).

While this does forces the user to call an additional function, failing
to do so should cause the protocol to fail (because it would depend on
uninitialised memory).  Sanitisers should also detect the issue.

Overall, I think this is a net improvement.