Chacha20 can derive keys after all
My dream is coming true, we'll be able to get rid of Blake2b. We'll need
a proper security reduction, but here's the gist of it:

We rely on the assumption that the HChacha20 of a shared secret is as
good as uniformly random if the private halves of the shared secret are
unknown to the attacker.  NaCl and Libsodium rely on the same assumption
for their crypto_box() API.

Instead of XOR'ing them together directly (which could be bad if two
shared secrets happen to be the same), we provide domain separation by
hashing the previous chaining key first. This makes sure we only XOR
independent variables.

Furthermore, the nonce used to hash the shared secret and the previous
chaining key are different, to sever any remaining algebraic relations.

Once we have a proper security analysis, we should be good enough for a
public beta (the only way I know of to get reviews).