Facilitate several encryption keys per handshake
As it is, the state machine only supports one encrypted key between two
exchanges. This restriction is incompatible with a number of Noise
patterns. We might want to extend it in the future.

An easy way to extend this would be to increment the Chacha counter to
get a new set of keys. Doing so however would expose an algebraic
relation between those keys and the next chaining key (same counter,
same nonce).  So we instead use a different nonce for the expansion
phase.