Implementation is now automatically generated
The monokex.c and monokex.h files are exclusively generated code, with
no human modification.  I'm pretty proud of the fact that I managed to
make it look like manually written code.  I've even managed to preserve
the comments.

We're now ready to sunset this repository, and integrate the code to
Monocypher.