Handshake revolution
We're changing the internals of the handshake.  Again.

Handling payloads turned out to be more difficult than anticipated.
Giving users only a payload key is a usability nightmare.  It forces
them to check the message and its payload separately, in addition to
encrypting and decrypting that payload themselves.  This means the API
must take care of the payloads itself.

Moreover, there was a potential weakness compared to Noise: earlier
payloads are not integrated to later transcripts.  Unencrypted payloads
would remain forever unauthenticated.

We need an overhaul of the API *and* of the internals.  This time we
need a proper hash, so HChacha20 is out.  We need Blake2b.  While we're
at it, we'll use Blake2b instead of Poly1305 for authentication.
Chacha20 will only be used to encrypt payloads.

The new scheme revolves around a mix_hash() function that is similar to
that of Noise's, only simpler.  It's basically a sponge construction,
which uses a full hash (Blake2b) instead of a simple permutation.  The
inputs are a chaining key and the data to absorb.  The output is 64
bytes: 32 for the next chaining key, 32 for authentication or
encryption.

  new_CK, K = mix_hash(old_CK, input)

The transcript and the X25519 exchanges are absorbed in the same way.
This effectively merges Noise's mixKey() and mixHash().  I believe this
is the simplest design so far.  Possibly the simplest possible.

Plus, it removes limitations on the protocol: we can now encrypt several
keys in a row, support preludes, even support pre-shared keys.  Noise
was right to go with a state machine after all.  Our state machine is
just simpler.