API revolution (monokex.*, test.c)
The handshake API has been completely reworked.  It is now closer to
what we can find in Noise.

- Sending and receiving functionalities are now separate.
- Support for payloads have been added (to be tested).
- Query functions have been added to enable generic network code.

Note 1
------

Querying the session keys causes the context to be wiped.  It makes it
impossible to query the remote static key afterwards.  Users must
therefore query the remote key first, *then* query the session keys.  We
could alleviate this inconvenience by not wiping the remote key and
status flags, but I'm not sure how wise this would be.

Note 2
------

The API let us get *two* session keys:

    void crypto_kex_get_session_keys(crypto_kex_ctx *ctx,
                                     uint8_t key1[32], // ignored if 0
                                     uint8_t key2[32]);// ignored if 0

Alternatives could look like this:

    void crypto_kex_get_secret_hash(crypto_kex_ctx *ctx,
                                    uint8_t key[64]);

    void crypto_kex_get_session_key(crypto_kex_ctx *ctx,
                                    uint8_t key[32]);

I'm not sure which is best.  The first is ideal for bidirectional
channels, where it doesn't matter if nonces get reuse across directions
(reusing a nonce in a single direction is still a disaster, mind you).
Or the user could use one key as a session key, and the second as a
unique identifier, for instance for channel binding.

The second is closest to the actual implementation: internally, we have
a 64 byte hash, so that's what we give the user.  Simpler, but perhaps
not as intuitive as it should be.

The third just gives a single session key.  It's the simplest, but it
wastes 32 random bytes.  Users who want more need to call Blake2b or
Chacha20 once more.  The performance loss, though negligible, feels
inelegant.