Increased hash collision resistance
Previously, the chaining hash was only 256 bits.  I mistakenly deemed
that enough.  The reality is, it only provides 2^128 bits of collision
resistance (because birthday attacks).

While this may be enough for most application, users wanting to use X448
(which aims for 2^256 security) instead of X25519 (which only aims for
2^128, but that's enough because of reasons), would not be happy.

Plus, even with X25519, collisions may still be the weakest link,
because attacks do not scale the same way on hashes and on elliptic
curves (if I understood correctly, hashes are more vulnerable to multi
collision attacks).

The performance cost of the fix is minimal: about 0.05%.