Put test vectors in main tests
Test vectors aren't really compatible with Elligator. While Monocypher
does use a deterministic scheme, the details of the particular scheme is
supposed to be unspecified.

However, once we have the context with the private ephemeral key pair
and the public Elligator representative, we can reconstruct everything
from specs, which is exactly what we do here.

Note that so far, we only test the happy path.  We now need to test that
corrupting messages leads to a clean failure.