Streamlined the state machine
Nadim Kobeissi pointed out that many hash steps could be simplified or
eliminated. Doing so simplifies the state machine, and makes it a bit
more efficient.

Note that now our chaining hash is only 256-bits, reducing collision
security to 128 bits. This is okay, because this is equivalent to
Curve25519 security.

Now we use 2 key derivation functions:

    h1, key = Chacha20_stream(h0, nonce=0)
    h1, tag = keyed_blake2b  (h0, input_key_material)

We could use blake2b in both cases, but I figured Chacha20 is a bit more
efficient on 32-bit machines and smaller.